+++ ceammc/extra/SoundTouch/pd/soundtouch_tilde.cpp
@@ -4,13 +4,11 @@
 #include <cmath>
 
 constexpr t_float DEF_PITCH = 0;
-static t_symbol* PROP_ANTIALIAS_LENGTH;
 
 SoundTouchExt::SoundTouchExt(const PdArgs& a)
     : SoundExternal(a)
-    , bypass_(nullptr)
-    , pitch_(nullptr)
     , pitch_value_(DEF_PITCH)
+    , drywet_(1)
 {
     createSignalOutlet();
     createInlet();
@@ -43,6 +41,16 @@ SoundTouchExt::SoundTouchExt(const PdArgs& a)
     bypass_ = new BoolProperty("@bypass", false);
     addProperty(bypass_);
 
+    {
+        auto dw = new FloatProperty("@drywet", 1);
+        dw->checkClosedRange(0, 1);
+        dw->setSuccessFn([dw, this](Property* p) {
+            drywet_.setTargetValue(dw->value());
+        });
+
+        addProperty(dw);
+    }
+
     initSoundTouch();
 }
 
@@ -55,22 +63,17 @@ void SoundTouchExt::processBlock(const t_sample** in, t_sample** out)
 
     const size_t bs = blockSize();
 
-#if PD_FLOATSIZE == 32
-    stouch_.putSamples(in[0], bs);
-    stouch_.receiveSamples(out[0], bs);
-#elif PD_FLOATSIZE == 64
     float fin[bs];
     float fout[bs];
 
-    for(size_t i = 0; i < bs; i++)
+    for (size_t i = 0; i < bs; i++)
         fin[i] = in[0][i];
 
     stouch_.putSamples(fin, bs);
     stouch_.receiveSamples(fout, bs);
 
-    for(size_t i = 0; i < bs; i++)
-        out[0][i] = fout[i];
-#endif
+    for (size_t i = 0; i < bs; i++)
+        out[0][i] = interpolate::linear<t_sample>(fout[i], in[0][i], drywet_());
 }
 
 void SoundTouchExt::onInlet(size_t, const AtomListView& lst)
@@ -95,8 +98,6 @@ void SoundTouchExt::initSoundTouch()
 
 extern "C" void soundtouch_tilde_setup()
 {
-    PROP_ANTIALIAS_LENGTH = gensym("@aalength");
-
     SoundExternalFactory<SoundTouchExt> obj("soundtouch~");
 
     LIB_DBG << "Soundtouch version: " << SOUNDTOUCH_VERSION;