+++ ceammc/extra/SoundTouch/pd/soundtouch_tilde.h
@@ -3,14 +3,16 @@
 
 #include "SoundTouch.h"
 
+#include "ceammc_signal.h"
 #include "ceammc_sound_external.h"
 using namespace ceammc;
 
 class SoundTouchExt : public SoundExternal {
     soundtouch::SoundTouch stouch_;
-    BoolProperty* bypass_;
-    Property* pitch_;
+    BoolProperty* bypass_ { nullptr };
+    Property* pitch_ { nullptr };
     t_float pitch_value_;
+    SmoothLinT<t_sample> drywet_;
 
 public:
     SoundTouchExt(const PdArgs& a);